Support detection already defined FK when other options are given
