+++ lib/safe-pg-migrations/plugins/idem_potent_statements.rb
@@ -27,7 +27,7 @@ module SafePgMigrations
     end
 
     def add_foreign_key(from_table, to_table, **options)
-      return super unless foreign_key_exists?(from_table, options.except(:validate).presence || to_table)
+      return super unless foreign_key_exists?(from_table, options.slice(:name, :column) || to_table)
 
       SafePgMigrations.say("/!\\ Foreign key '#{from_table}' -> '#{to_table}' already exists. Skipping statement.", true)
     end
