+++ test/safe_pg_migrations_test.rb
@@ -3,8 +3,6 @@
 require 'test_helper'
 
 class SafePgMigrationsTest < Minitest::Test
-  DUMMY_MIGRATION_VERSION = 8128
-
   def setup
     SafePgMigrations.instance_variable_set(:@config, nil)
     @connection = ActiveRecord::Base.connection
@@ -20,8 +18,6 @@ class SafePgMigrationsTest < Minitest::Test
     ActiveRecord::SchemaMigration.drop_table
     @connection.execute('SET statement_timeout TO 0')
     @connection.execute("SET lock_timeout TO '30s'")
-    @connection.drop_table(:messages, if_exists: true)
-    @connection.drop_table(:conversations, if_exists: true)
     @connection.drop_table(:users, if_exists: true)
     ActiveRecord::Migration.verbose = @verbose_was
   end
@@ -317,209 +313,6 @@ class SafePgMigrationsTest < Minitest::Test
     refute @connection.table_exists?(:users)
   end
 
-  def test_add_foreign_key_with_validate_explicitly_false
-    @connection.create_table(:users) { |t| t.string :email }
-    @connection.create_table(:messages) do |t|
-      t.string :message
-      t.bigint :user_id
-    end
-
-    @migration =
-      Class.new(ActiveRecord::Migration::Current) do
-        def change
-          add_foreign_key :messages, :users, validate: false
-        end
-      end.new
-
-    calls = record_calls(@connection, :execute) { run_migration }
-    assert_calls [
-      "SET statement_timeout TO '5s'",
-      'ALTER TABLE "messages" ADD CONSTRAINT "fk_rails_273a25a7a6" FOREIGN KEY ("user_id") ' \
-      'REFERENCES "users" ("id") NOT VALID',
-      "SET statement_timeout TO '70s'",
-    ], calls
-  end
-
-  def test_add_foreign_key
-    @connection.create_table(:users) { |t| t.string :email }
-    @connection.create_table(:messages) do |t|
-      t.string :message
-      t.bigint :user_id
-    end
-
-    @migration =
-      Class.new(ActiveRecord::Migration::Current) do
-        def change
-          add_foreign_key :messages, :users
-        end
-      end.new
-
-    calls = record_calls(@connection, :execute) { run_migration }
-    assert_calls [
-      "SET statement_timeout TO '5s'",
-      'ALTER TABLE "messages" ADD CONSTRAINT "fk_rails_273a25a7a6" FOREIGN KEY ("user_id") ' \
-      'REFERENCES "users" ("id") NOT VALID',
-      "SET statement_timeout TO '70s'",
-      'SET statement_timeout TO 0',
-      'ALTER TABLE "messages" VALIDATE CONSTRAINT "fk_rails_273a25a7a6"',
-      "SET statement_timeout TO '70s'",
-    ], calls
-  end
-
-  def test_add_foreign_key_idem_potent
-    @connection.create_table(:users) { |t| t.string :email }
-    @connection.create_table(:messages) do |t|
-      t.string :message
-      t.bigint :user_id
-    end
-
-    @migration =
-      Class.new(ActiveRecord::Migration::Current) do
-        def change
-          2.times { add_foreign_key :messages, :users }
-        end
-      end.new
-
-    execute_calls = nil
-    write_calls =
-        record_calls(@migration, :write) do
-          execute_calls = record_calls(@connection, :execute) { run_migration }
-        end
-    assert_calls [
-      "SET statement_timeout TO '5s'",
-      'ALTER TABLE "messages" ADD CONSTRAINT "fk_rails_273a25a7a6" FOREIGN KEY ("user_id") ' \
-      'REFERENCES "users" ("id") NOT VALID',
-      "SET statement_timeout TO '70s'",
-      'SET statement_timeout TO 0',
-      'ALTER TABLE "messages" VALIDATE CONSTRAINT "fk_rails_273a25a7a6"',
-      "SET statement_timeout TO '70s'",
-      "SET statement_timeout TO '5s'",
-      "SET statement_timeout TO '70s'",
-      "SET statement_timeout TO 0",
-      'ALTER TABLE "messages" VALIDATE CONSTRAINT "fk_rails_273a25a7a6"',
-      "SET statement_timeout TO '70s'",
-    ], execute_calls
-
-    assert_equal [
-      "== 8128 : migrating ===========================================================",
-      "-- add_foreign_key(:messages, :users)",
-      "-- add_foreign_key(:messages, :users)",
-      "   -> /!\\ Foreign key 'messages' -> 'users' already exists. Skipping statement."
-    ], write_calls.map(&:first).values_at(0,1,3,4)
-  end
-
-  def test_add_foreign_key_idem_potent_with_column_option
-    @connection.create_table(:users) { |t| t.string :email }
-    @connection.create_table(:messages) do |t|
-      t.string :message
-      t.bigint :author_id
-    end
-
-    @migration =
-      Class.new(ActiveRecord::Migration::Current) do
-        def change
-          2.times { add_foreign_key :messages, :users, column: :author_id }
-        end
-      end.new
-
-    execute_calls = nil
-    write_calls =
-        record_calls(@migration, :write) do
-          execute_calls = record_calls(@connection, :execute) { run_migration }
-        end
-
-    assert_calls [
-      "SET statement_timeout TO '5s'",
-      'ALTER TABLE "messages" ADD CONSTRAINT "fk_rails_995937c106" FOREIGN KEY ("author_id") ' \
-      'REFERENCES "users" ("id") NOT VALID',
-      "SET statement_timeout TO '70s'",
-      'SET statement_timeout TO 0',
-      'ALTER TABLE "messages" VALIDATE CONSTRAINT "fk_rails_995937c106"',
-      "SET statement_timeout TO '70s'",
-      "SET statement_timeout TO '5s'",
-      "SET statement_timeout TO '70s'",
-      "SET statement_timeout TO 0",
-      'ALTER TABLE "messages" VALIDATE CONSTRAINT "fk_rails_995937c106"',
-      "SET statement_timeout TO '70s'",
-    ], execute_calls
-
-      assert_equal [
-        "== 8128 : migrating ===========================================================",
-        "-- add_foreign_key(:messages, :users, {:column=>:author_id})",
-        "-- add_foreign_key(:messages, :users, {:column=>:author_id})",
-        "   -> /!\\ Foreign key 'messages' -> 'users' already exists. Skipping statement."
-      ], write_calls.map(&:first).values_at(0,1,3,4)
-  end
-
-  def test_add_foreign_key_idem_potent_different_tables
-    @connection.create_table(:users) { |t| t.string :email }
-    @connection.create_table(:conversations) { |t| t.string :subject }
-    @connection.create_table(:messages) do |t|
-      t.string :message
-      t.bigint :author_id
-      t.bigint :conversation_id
-    end
-
-    @migration =
-      Class.new(ActiveRecord::Migration::Current) do
-        def change
-          add_foreign_key :messages, :users, column: :author_id
-          add_foreign_key :messages, :conversations
-        end
-      end.new
-
-    execute_calls = nil
-    write_calls =
-        record_calls(@migration, :write) do
-          execute_calls = record_calls(@connection, :execute) { run_migration }
-        end
-
-    assert_calls [
-      "SET statement_timeout TO '5s'",
-      'ALTER TABLE "messages" ADD CONSTRAINT "fk_rails_995937c106" FOREIGN KEY ("author_id") ' \
-      'REFERENCES "users" ("id") NOT VALID',
-      "SET statement_timeout TO '70s'",
-      'SET statement_timeout TO 0',
-      'ALTER TABLE "messages" VALIDATE CONSTRAINT "fk_rails_995937c106"',
-      "SET statement_timeout TO '70s'",
-      "SET statement_timeout TO '5s'",
-      'ALTER TABLE "messages" ADD CONSTRAINT "fk_rails_7f927086d2" FOREIGN KEY ("conversation_id") ' \
-      'REFERENCES "conversations" ("id") NOT VALID',
-      "SET statement_timeout TO '70s'",
-      "SET statement_timeout TO 0",
-      'ALTER TABLE "messages" VALIDATE CONSTRAINT "fk_rails_7f927086d2"',
-      "SET statement_timeout TO '70s'",
-    ], execute_calls
-
-    assert_equal [
-      "== 8128 : migrating ===========================================================",
-      "-- add_foreign_key(:messages, :users, {:column=>:author_id})",
-      "-- add_foreign_key(:messages, :conversations)",
-    ], write_calls.map(&:first).values_at(0,1,3)
-  end
-
-  def test_add_foreign_key_with_validation
-    @connection.create_table(:users) { |t| t.string :email }
-    @connection.create_table(:messages) do |t|
-      t.string :message
-      t.bigint :user_id
-    end
-
-    @migration =
-      Class.new(ActiveRecord::Migration::Current) do
-        def change
-          add_foreign_key :messages, :users, validate: true
-        end
-      end.new
-
-    calls = record_calls(@connection, :execute) { run_migration }
-    assert_calls [
-      "SET statement_timeout TO '5s'",
-      'ALTER TABLE "messages" ADD CONSTRAINT "fk_rails_273a25a7a6" FOREIGN KEY ("user_id") REFERENCES "users" ("id")',
-      "SET statement_timeout TO '70s'",
-    ], calls
-  end
-
   def test_add_index
     @connection.create_table(:users) { |t| t.string :email }
     @migration =
@@ -709,53 +502,5 @@ class SafePgMigrationsTest < Minitest::Test
       assert_match('SELECT version()', logs[3])
       assert_match("SET lock_timeout TO '70s'", logs[4])
     end
-  end
-
-  def run_migration(direction = :up)
-    @migration.version = DUMMY_MIGRATION_VERSION
-    ActiveRecord::Migrator.new(direction, [@migration]).migrate
-  end
-
-  def assert_calls(expected, actual)
-    assert_equal [
-      "SET lock_timeout TO '5s'",
-      *expected,
-      "SET lock_timeout TO '70s'",
-    ], actual[0...-4].map(&:first).map(&:squish)
-  end
-
-  # Records method calls on an object. Behaves like a test spy.
-  #
-  # Example usage:
-  #
-  #   record_calls(foo, :bar) { foo.bar(1, 2); foo.bar(3, 4) }
-  #
-  # Example return:
-  #
-  #   [[1, 2], [3, 4]]
-  #
-  def record_calls(object, method)
-    calls = []
-    recorder =
-      lambda {
-        object.stubs(method).with do |*args|
-          calls << args
-          # Temporarily unstub the method so that we can call the original method.
-          object.unstub(method)
-          begin
-            # Call the original method.
-            object.send(method, *args)
-          ensure
-            # Register the recorder again.
-            recorder.call
-          end
-          true
-        end
-      }
-    recorder.call
-    yield
-    calls
-  ensure
-    object.unstub(method)
   end
 end
